@@ -17,6 +17,9 @@
 
 @inject AuthenticationStateProvider AuthenticationStateProvider
 @inject IJSRuntime JSRuntime
+@inject ErpService ObjErpService
+@inject ErpApiService ObjErpApiService
+
 @inject ReportService UserlogService
 @inject InboundService ObjInboundService
 @inject UserServices OpjUserService
@@ -63,7 +66,7 @@ else
               Filter="new Func<Inb_Goodreceive_Go,bool>(FilterFunc)" @bind-SelectedItem="selectedItem">
         <ToolBarContent>
 
-            <MudToolBarSpacer />
+            <MudSpacer />
             <MudTextField @bind-Value="searchString"
                           Placeholder=@L["Search"]
                           Adornment="Adornment.Start"
@@ -88,6 +91,7 @@ else
             <col />
             <col />
             <col />
+            <col />
         </ColGroup>
         <HeaderContent>
             <MudTh><MudTableSortLabel SortBy="new Func<Inb_Goodreceive_Go, object>(x=>x.Created)">@L["GRDatetime"]</MudTableSortLabel></MudTh>
@@ -100,9 +104,13 @@ else
             <MudTh><MudTableSortLabel SortBy="new Func<Inb_Goodreceive_Go, object>(x=>x.Weight)">@L["Weight"]</MudTableSortLabel></MudTh>
             <MudTh><MudTableSortLabel SortBy="new Func<Inb_Goodreceive_Go, object>(x=>x.Docno)">@L["LotNo"]</MudTableSortLabel></MudTh>
             <MudTh><MudTableSortLabel SortBy="new Func<Inb_Goodreceive_Go, object>(x=>x.Docnote)">@L["BatchNo"]</MudTableSortLabel></MudTh>
+            <MudTh>@L["Adjust"]</MudTh>
         </HeaderContent>
         <GroupHeaderTemplate>
-            <MudTh Class="mud-table-cell-custom-group" colspan="10">@($"{context.GroupName}: {context.Key} ") Total: @string.Format(VarGlobals.FormatN0, @context.Items.Sum((e) => e.Quantity))</MudTh>
+            <MudTh Class="mud-table-cell-custom-group" colspan="10">
+                @($"{context.GroupName}: {context.Key} ") Total: @string.Format(VarGlobals.FormatN0, @context.Items.Sum((e) => e.Quantity))
+                <MudFab  @onclick="@(()=>DeleteMapPallet(context.Key.ToString()))"  Color="Color.Error" Icon="@Icons.Material.Filled.AutoDelete" Size="Size.Small" IconSize="Size.Small" />
+            </MudTh>
         </GroupHeaderTemplate>
         <RowTemplate>
             <MudTd DataLabel="@L["GRDatetime"]">@Convert.ToDateTime(context.Created).ToString(VarGlobals.FormatDT)</MudTd>
@@ -115,7 +123,9 @@ else
             <MudTd DataLabel="@L["Weight"]" Style="text-align: right">@string.Format(VarGlobals.FormatN2, context.Weight)</MudTd>
             <MudTd DataLabel="@L["LotNo"]">@context.Docno</MudTd>
             <MudTd DataLabel="@L["BatchNo"]">@context.Docnote</MudTd>
-
+            <MudTd DataLabel="@L["Adjust"]">
+                <MudFab @onclick="@(()=>DeleteMap(@context.Pallteno, @context.Pallettag))" Color="Color.Secondary" Icon="@Icons.Material.Filled.Delete" Size="Size.Small" IconSize="Size.Small" />
+            </MudTd>
         </RowTemplate>
         <GroupFooterTemplate>
             <MudTh Class="mud-table-cell-custom-group mud-table-cell-custom-group-footer" colspan="10">Total Quantity: @string.Format(VarGlobals.FormatN0, @context.Items.Sum((e) => e.Quantity)) </MudTh>
@@ -240,7 +250,7 @@ else
         try
         {
             bool? result = await CheckRoleReport();
-            if (result==true)
+            if (result == true)
             {
                 IbpOrderQueueRptExcel excelReport = new IbpOrderQueueRptExcel();
                 List<Inb_Goodreceive_Go> listExcelReport = new List<Inb_Goodreceive_Go>();
@@ -279,6 +289,42 @@ else
         {
             var exmsg = e.Message.ToString();
         }
+    }
+
+
+    private async Task DeleteMap(string pallet, string pack)
+    {
+        bool? result = await CheckRoleApprove();
+
+        if (result == true)
+        {
+            ObjInboundService.CancelReceivingOrdersBypack(pallet, pack);
+
+            this.InsertAuditrial("Cancel ReceivingOrder SeqNO : " + pack + " From Pallet : " + pallet, pageheader);
+
+            Elements = ObjInboundService.GetAllInbGoodreceiveGos();
+        }
+
+        this.StateHasChanged();
+
+    }
+
+
+    private async Task DeleteMapPallet(string pallet)
+    {
+        bool? result = await CheckRoleApprove();
+
+        if (result == true)
+        {
+            ObjInboundService.CancelReceivingOrdersByPallet(pallet);
+
+            this.InsertAuditrial("Cancel ReceivingOrders From Pallet : " + pallet, pageheader);
+
+            Elements = ObjInboundService.GetAllInbGoodreceiveGos();
+        }
+
+        this.StateHasChanged();
+
     }
 
     #region "Role"
