@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Configuration;
 using System.Data;
+using System.Data.SqlClient;
 using Npgsql;
 using System.Linq;
 using System.Threading.Tasks;
@@ -16,6 +17,7 @@ namespace GoWMS.Server.Data
     public class InbDAL
     {
         readonly private string connectionString = ConnGlobals.GetConnLocalDBPG();
+        readonly private string connectionStringSQL = ConnGlobals.GetConnDBSQL();
 
         public IEnumerable<Inb_Goodreceipt_Go> GetAllInbGoodreceiptGo()
         {
@@ -23,7 +25,7 @@ namespace GoWMS.Server.Data
             using (NpgsqlConnection con = new NpgsqlConnection(connectionString))
             {
                 NpgsqlCommand cmd = new NpgsqlCommand("select * " +
-                    "from wms.inb_goodreceipt_go  " +
+                    "from wms.inb_goodreceipt_go " +
                     "order by efidx", con)
                 {
                     CommandType = CommandType.Text
@@ -72,7 +74,7 @@ namespace GoWMS.Server.Data
             using (NpgsqlConnection con = new NpgsqlConnection(connectionString))
             {
                 NpgsqlCommand cmd = new NpgsqlCommand("select * " +
-                    "from wms.inb_goodreceive_go  " +
+                    "from wms.inb_goodreceive_go " +
                     "order by efidx", con)
                 {
                     CommandType = CommandType.Text
@@ -250,14 +252,33 @@ namespace GoWMS.Server.Data
         public async Task<Int64> GetSumOrderAllInbGoodreceiptGo()
         {
             Int64 lRet = 0;
-            using (NpgsqlConnection con = new NpgsqlConnection(connectionString))
+            using (SqlConnection con = new SqlConnection(connectionStringSQL))
             {
-                NpgsqlCommand cmd = new NpgsqlCommand("select count(*) " +
-                    "from wms.inb_goodreceive_go " +
-                    "where pallteno in (select palletno from wms.inb_putaway_go)", con)
+                StringBuilder sqlQurey = new StringBuilder();
+
+                sqlQurey.AppendLine("select count(*) ");
+                sqlQurey.AppendLine("from [dbo].[wms_trans] t ");
+                sqlQurey.AppendLine("inner join [dbo].[v_itemmaster] i ");
+                sqlQurey.AppendLine("on t.[item]=i.[item] ");
+                sqlQurey.AppendLine("where 1=1");
+                sqlQurey.AppendLine("and t.[is_req]=0");
+                sqlQurey.AppendLine("and t.[stat]=0");
+                sqlQurey.AppendLine("and t.[flag]=0");
+                sqlQurey.AppendLine("and t.[unit_key] in ('01')");
+                sqlQurey.AppendLine(";");
+
+                SqlCommand cmd = new SqlCommand(sqlQurey.ToString(), con)
                 {
                     CommandType = CommandType.Text
                 };
+
+                //NpgsqlCommand cmd = new NpgsqlCommand("select count(*) " +
+                //   "from wms.inb_goodreceive_go " +
+                //   "where pallteno in (select palletno from wms.inb_putaway_go)", con)
+                //{
+                //    CommandType = CommandType.Text
+                //};
+
                 con.Open();
 
                 lRet= Convert.ToInt64(cmd.ExecuteScalar());
@@ -271,14 +292,28 @@ namespace GoWMS.Server.Data
         public async Task<Int64> GetSumPalletAllInbGoodreceiptGo()
         {
             Int64 lRet = 0;
-            using (NpgsqlConnection con = new NpgsqlConnection(connectionString))
+            using (SqlConnection con = new SqlConnection(connectionStringSQL))
             {
-                NpgsqlCommand cmd = new NpgsqlCommand("select count(*) " +
-                    "from wms.inb_putaway_go " +
-                    "where efstatus=0 ", con)
+                StringBuilder sqlQurey = new StringBuilder();
+
+                sqlQurey.AppendLine("select count(lpncode)");
+                sqlQurey.AppendLine("from wcs.tas_works");
+                sqlQurey.AppendLine("where work_code='01'");
+                sqlQurey.AppendLine("and work_status='AVL'");
+                sqlQurey.AppendLine(";");
+
+                SqlCommand cmd = new SqlCommand(sqlQurey.ToString(), con)
                 {
                     CommandType = CommandType.Text
                 };
+
+                //NpgsqlCommand cmd = new NpgsqlCommand("select count(*) " +
+                //    "from wms.inb_putaway_go " +
+                //    "where efstatus=0 ", con)
+                //{
+                //    CommandType = CommandType.Text
+                //};
+
                 con.Open();
                 lRet = Convert.ToInt64(cmd.ExecuteScalar());
 
@@ -292,9 +327,9 @@ namespace GoWMS.Server.Data
         public async Task<Int64> GetSumPalletAllOubGoodPickingGo()
         {
             Int64 lRet = 0;
-            using (NpgsqlConnection con = new NpgsqlConnection(connectionString))
+            using (SqlConnection con = new SqlConnection(connectionStringSQL))
             {
-                NpgsqlCommand cmd = new NpgsqlCommand("select count(lpncode) " +
+                SqlCommand cmd = new SqlCommand("select count(lpncode) " +
                     "from wcs.tas_works " +
                     "where work_code='05' " +
                     "and work_status='AVL' ", con)
@@ -313,10 +348,10 @@ namespace GoWMS.Server.Data
         public async Task<Int64> GetSumOrderAllOubGoodPickingGo()
         {
             Int64 lRet = 0;
-            using (NpgsqlConnection con = new NpgsqlConnection(connectionString))
+            using (SqlConnection con = new SqlConnection(connectionStringSQL))
             {
-                NpgsqlCommand cmd = new NpgsqlCommand("select count(*) " +
-                    "from public.sap_storeout " +
+                SqlCommand cmd = new SqlCommand("select count(*) " +
+                    "from dbo.sap_storeout " +
                     "where status<3 ", con)
                 {
                     CommandType = CommandType.Text
@@ -331,5 +366,86 @@ namespace GoWMS.Server.Data
         }
 
 
+        public bool CancelReceivingOrdersBypack(string pallet, string pack)
+        {
+            bool bret=false;
+            try
+            {
+                using NpgsqlConnection con = new NpgsqlConnection(connectionString);
+                StringBuilder sql = new StringBuilder();
+
+                sql.AppendLine("Delete From wms.inb_goodreceive_go");
+                sql.AppendLine("Where pallettag = @Pack ");
+                sql.AppendLine("and pallteno = @Pallet");
+                sql.AppendLine(";");
+
+                sql.AppendLine("Delete From public.sap_storein");
+                sql.AppendLine("Where su_no = @su_no ");
+                sql.AppendLine("and sap_su = @sap_su");
+                sql.AppendLine(";");
+
+                NpgsqlCommand cmd = new NpgsqlCommand(sql.ToString(), con)
+                {
+                    CommandType = CommandType.Text
+                };
+
+                cmd.Parameters.AddWithValue("@Pallet", pallet);
+                cmd.Parameters.AddWithValue("@Pack", pack);
+
+                cmd.Parameters.AddWithValue("@sap_su", pallet);
+                cmd.Parameters.AddWithValue("@su_no", pack);
+                con.Open();
+                cmd.ExecuteNonQuery();
+                con.Close();
+                bret = true;
+            }
+            catch
+            {
+
+            }
+            return bret;
+        }
+
+
+        public bool CancelReceivingOrdersByPallet(string pallet)
+        {
+            bool bret = false;
+            try
+            {
+                using NpgsqlConnection con = new NpgsqlConnection(connectionString);
+                StringBuilder sql = new StringBuilder();
+
+                sql.AppendLine("Delete From wms.inb_goodreceive_go");
+                sql.AppendLine("Where pallteno = @pallteno");
+                sql.AppendLine(";");
+
+                sql.AppendLine("Delete From public.sap_storein");
+                sql.AppendLine("Where sap_su = @sap_su");
+                sql.AppendLine(";");
+
+                sql.AppendLine("Delete From wcs.tas_works");
+                sql.AppendLine("Where lpncode = @lpncode ");
+                sql.AppendLine(";");
+
+                NpgsqlCommand cmd = new NpgsqlCommand(sql.ToString(), con)
+                {
+                    CommandType = CommandType.Text
+                };
+
+                cmd.Parameters.AddWithValue("@pallteno", pallet);
+                cmd.Parameters.AddWithValue("@sap_su", pallet);
+                cmd.Parameters.AddWithValue("@lpncode", pallet);
+
+                con.Open();
+                cmd.ExecuteNonQuery();
+                con.Close();
+                bret = true;
+            }
+            catch
+            {
+
+            }
+            return bret;
+        }
     }
 }
