@@ -8,7 +8,7 @@ namespace GoWMS.Server.Models.Wcs
     public class Vmachine_command
     {
         public string Mccode { get; set; }
-        public Int32? St_no { get; set; }
+        public Int32 St_no { get; set; }
         public string St_desc_th { get; set; }
         public string Sg_no { get; set; }
         public string Sg_desc_th { get; set; }
