@@ -15,7 +15,7 @@
             </MudHidden>
             <MudText Typo="Typo.h6" Inline="true" Class="mr-2">#</MudText>
             <MudText Typo="Typo.h6">chat</MudText>
-            <MudToolBarSpacer />
+            <MudSpacer />
             <MudIconButton Icon="@Icons.Material.Outlined.Notifications" />
             <MudIconButton Icon="@Icons.Material.Outlined.PushPin" />
             <MudHidden Breakpoint="Breakpoint.MdAndUp">
