@@ -20,7 +20,7 @@
     <MudPaper Elevation="25" Class="py-4 flex-grow-1">
         <MudToolBar Dense="true" DisableGutters="true">
             <MudCheckBox T="bool" Label="Select All" Class="ml-2" />
-            <MudAppBarSpacer />
+            <MudSpacer />
             <MudIconButton Icon="@Icons.Material.Outlined.NavigateBefore" Class="mx-2" />
             <MudIconButton Icon="@Icons.Material.Outlined.NavigateNext" Class="mr-4" />
         </MudToolBar>
