@@ -61,7 +61,7 @@
                   Hover="@hover"
                   Bordered="@bordered" Striped="@striped" Filter="new Func<AsrsPerformance, bool>(FilterFunc)" Style="width:100%;">
             <ToolBarContent>
-                <MudToolBarSpacer />
+                <MudSpacer />
                 <MudTextField @bind-Value="searchString"
                               Placeholder=@L["Search"] Adornment="Adornment.Start"
                               AdornmentIcon="@Icons.Material.Filled.Search"
