@@ -66,13 +66,13 @@ else
 
         <HeaderContent>
             <MudTh><MudTableSortLabel SortBy="new Func<Vmachine, object>(x=>x.Mccode)">@L["MCCODE"]</MudTableSortLabel></MudTh>
-            <MudTh><MudTableSortLabel SortBy="new Func<Vmachine, object>(x=>x.Information)">@L["INFO"]</MudTableSortLabel></MudTh>
+            @*<MudTh><MudTableSortLabel SortBy="new Func<Vmachine, object>(x=>x.Information)">@L["INFO"]</MudTableSortLabel></MudTh>*@
             <MudTh><MudTableSortLabel SortBy="new Func<Vmachine, object>(x=>x.St_no)">@L["STATUS"]</MudTableSortLabel></MudTh>
             <MudTh>@L["ACTION"]</MudTh>
         </HeaderContent>
         <RowTemplate>
             <MudTd DataLabel=@L["MCCODE"]>@context.Mccode</MudTd>
-            <MudTd DataLabel=@L["INFO"]>@context.Information</MudTd>
+            @*<MudTd DataLabel=@L["INFO"]>@context.Information</MudTd>*@
             <MudTd DataLabel=@L["STATUS"]>
                 @{if (@context.Is_alert == true && @context.Mccode != "")
                         {
@@ -103,8 +103,10 @@ else
 
                 @if (@context.Is_cmd == true)
                     {
-                    <MudFab  @onclick="@(()=>OpenDialog(@context.Mccode))" Color="Color.Error" Icon="@Icons.Material.Filled.MenuOpen" Size="Size.Small" IconSize="Size.Small" />
+                    <MudFab  @onclick="@(()=>OpenDialog(@context.Mccode))" Color="Color.Error" Icon="@Icons.Filled.EditRoad" Size="Size.Small" IconSize="Size.Small" />
                     }
+                 <MudFab  Color="Color.Info" Icon="@Icons.Filled.Info" Size="Size.Small" IconSize="Size.Small" />
+
             </MudTd>
         </RowTemplate>
         <PagerContent>
@@ -172,12 +174,12 @@ else
 
         ConstElements = await Task.Run(() => ObjWcsService.GetConstance());
 
-        var result1 = ConstElements.Where(i => i.Set_code == "HMI001").FirstOrDefault(); // = 0
-        var result2 = ConstElements.Where(i => i.Set_code == "HMI002").FirstOrDefault(); // = 0
+        //var result1 = ConstElements.Where(i => i.Set_code == "HMI001").FirstOrDefault(); // = 0
+        //var result2 = ConstElements.Where(i => i.Set_code == "HMI002").FirstOrDefault(); // = 0
 
 
-        UrlHMI1 = result1.Val_vhr;
-        UrlHMI2 = result2.Val_vhr;
+        //UrlHMI1 = result1.Val_vhr;
+        //UrlHMI2 = result2.Val_vhr;
 
         this.ExecuteStatus();
 
