@@ -38,7 +38,7 @@ else
               Filter="new Func<Helpdesk,bool>(FilterFunc)" @bind-SelectedItem="selectedItem"
               SortLabel=@L["SortBy"]>
         <ToolBarContent>
-            <MudToolBarSpacer />
+            <MudSpacer />
             <MudTextField @bind-Value="searchString"
                           Placeholder=@L["Search"] Adornment="Adornment.Start"
                           AdornmentIcon="@Icons.Material.Filled.Search"
