@@ -84,7 +84,7 @@
                           Filter="new Func<Oub_CustomerReturn,bool>(FilterFunc)" @bind-SelectedItems="selectedItems">
                     <ToolBarContent>
 
-                        <MudToolBarSpacer />
+                        <MudSpacer />
                         <MudTextField @bind-Value="searchString"
                                       Placeholder=@L["Search"]
                                       Adornment="Adornment.Start"
