@@ -44,7 +44,7 @@ else
               SortLabel=@L["SortBy"]>
         <ToolBarContent>
 
-            <MudToolBarSpacer />
+            <MudSpacer />
             <MudTextField @bind-Value="searchString"
                           Placeholder=@L["Search"]
                           Adornment="Adornment.Start"
