@@ -42,7 +42,7 @@ else
               RowsPerPage="5">
         <ToolBarContent>
             <MudText Typo="Typo.h6">List Of Material Need</MudText>
-            <MudToolBarSpacer />
+            <MudSpacer />
             <MudTextField @bind-Value="searchStringLON" Placeholder="Search" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.Search" IconSize="Size.Medium" Class="mt-0"></MudTextField>
         </ToolBarContent>
 
@@ -79,7 +79,7 @@ else
 
         <ToolBarContent>
             <MudText Typo="Typo.h6">Reserved Materials</MudText>
-            <MudToolBarSpacer />
+            <MudSpacer />
             <MudTextField @bind-Value="searchStringRES" Placeholder="Search" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.Search" IconSize="Size.Medium" Class="mt-0"></MudTextField>
         </ToolBarContent>
 
@@ -116,7 +116,7 @@ else
                   RowsPerPage="5">
             <ToolBarContent>
                 <MudText Typo="Typo.h6">List Of Material Need</MudText>
-                <MudToolBarSpacer />
+                <MudSpacer />
                 <MudTextField @bind-Value="searchStringLON" Placeholder="Search" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.Search" IconSize="Size.Medium" Class="mt-0"></MudTextField>
             </ToolBarContent>
             <HeaderContent>
@@ -141,7 +141,7 @@ else
                   RowsPerPage="5">
             <ToolBarContent>
                 <MudText Typo="Typo.h6">Reserved Materials</MudText>
-                <MudToolBarSpacer />
+                <MudSpacer />
                 <MudTextField @bind-Value="searchStringRES" Placeholder="Search" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.Search" IconSize="Size.Medium" Class="mt-0"></MudTextField>
             </ToolBarContent>
 
