@@ -45,7 +45,7 @@ else
                   RowsPerPage="5">
             <ToolBarContent>
                 <MudText Typo="Typo.h6">List Of Material Need</MudText>
-                <MudToolBarSpacer />
+                <MudSpacer />
                 <MudTextField @bind-Value="searchStringLON" Placeholder="Search" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.Search" IconSize="Size.Medium" Class="mt-0"></MudTextField>
             </ToolBarContent>
 
@@ -82,7 +82,7 @@ else
 
             <ToolBarContent>
                 <MudText Typo="Typo.h6">Reserved Materials</MudText>
-                <MudToolBarSpacer />
+                <MudSpacer />
                 <MudTextField @bind-Value="searchStringRES" Placeholder="Search" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.Search" IconSize="Size.Medium" Class="mt-0"></MudTextField>
             </ToolBarContent>
 
@@ -119,7 +119,7 @@ else
                   RowsPerPage="5">
             <ToolBarContent>
                 <MudText Typo="Typo.h6">List Of Material Need</MudText>
-                <MudToolBarSpacer />
+                <MudSpacer />
                 <MudTextField @bind-Value="searchStringLON" Placeholder="Search" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.Search" IconSize="Size.Medium" Class="mt-0"></MudTextField>
             </ToolBarContent>
             <HeaderContent>
@@ -144,7 +144,7 @@ else
                   RowsPerPage="5">
             <ToolBarContent>
                 <MudText Typo="Typo.h6">Reserved Materials</MudText>
-                <MudToolBarSpacer />
+                <MudSpacer />
                 <MudTextField @bind-Value="searchStringRES" Placeholder="Search" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.Search" IconSize="Size.Medium" Class="mt-0"></MudTextField>
             </ToolBarContent>
 
