@@ -18,7 +18,7 @@
 
 <MudText Typo="Typo.h4" GutterBottom="true">Sign In to GoWMS</MudText>
 
-<MudText>Don't have an account? <MudLink Href="/pages/authentication/register" Color="Color.Error">Sign Up</MudLink></MudText>
+@* <MudText>Don't have an account? <MudLink Href="/pages/authentication/register" Color="Color.Error">Sign Up</MudLink></MudText>*@
 
 <MudTextField T="string" @bind-Value="@Userword" Label="UserID" Variant="Variant.Outlined" Class="my-6"></MudTextField>
 
@@ -27,7 +27,7 @@
 <br />
 
 
-<MudButton Variant="Variant.Filled" Color="Color.Error" Size="Size.Large" FullWidth="true" OnClick="@btnLogin_Click">Sign In</MudButton>
+<MudButton Variant="Variant.Filled" Color="Color.Primary" Size="Size.Large" FullWidth="true" OnClick="@btnLogin_Click">Sign In</MudButton>
 
 
 
