@@ -62,7 +62,7 @@
                   Bordered="@bordered" Striped="@striped" Filter="new Func<Rpt_Ejectgate, bool>(FilterFunc)" Style="width:100%;">
             <ToolBarContent>
 
-                <MudToolBarSpacer />
+                <MudSpacer />
                 <MudTextField @bind-Value="searchString" Placeholder="Search" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.Search" IconSize="Size.Medium" Class="mt-0"></MudTextField>
             </ToolBarContent>
             <HeaderContent>
