@@ -54,7 +54,7 @@ else
                   Dense="@dense" Hover="@hover" Bordered="@bordered" Striped="@striped"
                   Filter="new Func<Inv_Stock_GoInfo,bool>(FilterFunc)" @bind-SelectedItem="selectedItem">
             <ToolBarContent>
-                <MudToolBarSpacer />
+                <MudSpacer />
                 <MudTextField @bind-Value="searchString"
                               Placeholder=@L["Search"]
                               Adornment="Adornment.Start"
@@ -68,7 +68,6 @@ else
                 <MudTh><MudTableSortLabel SortBy="new Func<Inv_Stock_GoInfo, object>(x=>x.Docno)">@L["Lot"]</MudTableSortLabel></MudTh>
                 <MudTh><MudTableSortLabel SortBy="new Func<Inv_Stock_GoInfo, object>(x=>x.Pallettag)">@L["SEQ"]</MudTableSortLabel></MudTh>
                 <MudTh><MudTableSortLabel SortBy="new Func<Inv_Stock_GoInfo, object>(x=>x.Pallteno)">@L["PalletGo"]</MudTableSortLabel></MudTh>
-                <MudTh><MudTableSortLabel SortBy="new Func<Inv_Stock_GoInfo, object>(x=>x.Itembar)">@L["PalletErp"]</MudTableSortLabel></MudTh>
                 <MudTh><MudTableSortLabel SortBy="new Func<Inv_Stock_GoInfo, object>(x=>x.Storagearea)">@L["Area"]</MudTableSortLabel></MudTh>
                 <MudTh><MudTableSortLabel SortBy="new Func<Inv_Stock_GoInfo, object>(x=>x.Storagebin)">@L["Location"]</MudTableSortLabel></MudTh>
             </HeaderContent>
@@ -79,7 +78,6 @@ else
                 <MudTd DataLabel=@L["Lot"]>@context.Docno</MudTd>
                 <MudTd DataLabel=@L["SEQ"]>@context.Pallettag</MudTd>
                 <MudTd DataLabel=@L["PalletGo"]>@context.Pallteno</MudTd>
-                <MudTd DataLabel=@L["PalletErp"]>@context.Itembar</MudTd>
                 <MudTd DataLabel=@L["Area"]>@context.Storagearea</MudTd>
                 <MudTd DataLabel=@L["Location"]>@context.Storagebin</MudTd>
             </RowTemplate>
@@ -97,7 +95,7 @@ else
                   Dense="@dense" Hover="@hover" Bordered="@bordered" Striped="@striped"
                   Filter="new Func<InvStockSum,bool>(FilterFunc_)" @bind-SelectedItem="selectedItem_">
             <ToolBarContent>
-                <MudToolBarSpacer />
+                <MudSpacer />
                 <MudTextField @bind-Value="searchString"
                               Placeholder=@L["Search"]
                               Adornment="Adornment.Start"
