@@ -64,7 +64,7 @@ else
               Filter="new Func<InvStockSumByCus,bool>(FilterFunc)" @bind-SelectedItem="selectedItem">
         <ToolBarContent>
 
-            <MudToolBarSpacer />
+            <MudSpacer />
             <MudTextField @bind-Value="searchString"
                           Placeholder=@L["Search"]
                           Adornment="Adornment.Start"
