@@ -42,7 +42,7 @@ else
               HorizontalScrollbar="true"
               Filter="new Func<DELIVERY_NOTE_ITEMS,bool>(FilterFunc)" @bind-SelectedItem="selectedItem">
         <ToolBarContent>
-            <MudToolBarSpacer />
+            <MudSpacer />
             <MudTextField @bind-Value="searchString"
                           Placeholder=@L["Search"]
                           Adornment="Adornment.Start"
