@@ -14,7 +14,7 @@
 
         </MudPaper>
 
-        <MudAlert Severity="Severity.Error" Class="mt-8 mud-width-full" Style="max-width:500px;" Elevation="25" >No authentication is required, click sign in.</MudAlert>
+      @*  <MudAlert Severity="Severity.Error" Class="mt-8 mud-width-full" Style="max-width:500px;" Elevation="25" >No authentication is required, click sign in.</MudAlert>*@
     </div>
 </MudContainer>
 
