@@ -41,7 +41,7 @@
 
 
         </MudHidden>
-        <MudAppBarSpacer />
+        <MudSpacer />
         <ChooseLanguage />
 
         <MudIconButton Icon="@Icons.Material.Filled.Brightness4" Color="Color.Inherit" OnClick="@((e) => DarkMode())" />
@@ -134,9 +134,9 @@
         Palette = new Palette()
         {
             Primary = "#1E88E5",
-            AppbarBackground = "#D32F2F", //AppbarBackground = "#1E88E5",
+            AppbarBackground = "#0066ff", //AppbarBackground = "#1E88E5",
             Background = Colors.Grey.Lighten5,
-            DrawerBackground = "#FFF",
+            DrawerBackground = "#f2f2f2",
             DrawerText = "rgba(0,0,0, 0.7)",
             Success = "#007E33"
         },
