@@ -4,16 +4,16 @@
 <MudDivider Class="mb-1" />
 <MudNavMenu>
 
-    <MudNavLink Href="/gowms" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Dashboard">@L["Dashboard"]</MudNavLink>
+    <MudNavLink Href="/gowms" IconColor="Color.Primary" Icon="@Icons.Material.Filled.Dashboard">@L["Dashboard"]</MudNavLink>
 
-    <MudNavGroup Title=@L["1.Inbound"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Inbox">
+    <MudNavGroup Title=@L["1.Inbound"] IconColor="Color.Primary" Icon="@Icons.Material.Filled.Inbox">
         <MudNavLink Href="/1/1_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Input">@L["ReceivingOrder"] </MudNavLink>
         <MudNavLink Href="/1/1_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Input">@L["GoodReceive"]</MudNavLink>
         <MudNavLink Href="/1/1_3" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Input">@L["MasterQueuing"]</MudNavLink>
         <MudNavLink Href="/1/1_4" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Input">@L["Return"]</MudNavLink>
     </MudNavGroup>
 
-    <MudNavGroup Title=@L["2.Warehouse"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Store">
+    <MudNavGroup Title=@L["2.Warehouse"] IconColor="Color.Primary" Icon="@Icons.Material.Filled.Store">
         <MudNavLink Href="/2/2_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Storage">@L["StockList"]</MudNavLink>
         <MudNavLink Href="/2/2_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Storage">@L["Capacity"]</MudNavLink>
         <MudNavLink Href="/2/2_3" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Storage">@L["Location"]</MudNavLink>
@@ -21,20 +21,20 @@
         <MudNavLink Href="/2/2_5" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Storage">@L["Counting"]</MudNavLink>
     </MudNavGroup>
 
-    <MudNavGroup Title=@L["3.Outbound"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Outbox">
+    <MudNavGroup Title=@L["3.Outbound"] IconColor="Color.Primary" Icon="@Icons.Material.Filled.Outbox">
         <MudNavLink Href="/3/3_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Outbox">@L["OrderBooking"]</MudNavLink>
         <MudNavLink Href="/3/3_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Outbox">@L["OrderPicking"]</MudNavLink>
         <MudNavLink Href="/3/3_3" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Outbox">@L["QueuePicking"]</MudNavLink>
         <MudNavLink Href="/3/3_4" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Outbox">@L["Unplanned"]</MudNavLink>
     </MudNavGroup>
 
-    <MudNavGroup Title=@L["4.Automation"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Android">
+    <MudNavGroup Title=@L["4.Automation"] IconColor="Color.Primary" Icon="@Icons.Material.Filled.Android">
         <MudNavLink Href="/4/4_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Android">@L["4.1.ASRS-STATUS"]</MudNavLink>
         <MudNavLink Href="/4/4_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Android">@L["4.2.ASRS-QUEUE"]</MudNavLink>
         <MudNavLink Href="/4/4_3" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Android">@L["4.3.RGV-QUEUE"]</MudNavLink>
         <MudNavGroup Title=@L["4.4-RGV-SURVEY"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Android">
             <MudNavLink Href="/4/4_4_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Android">@L["4.4.1.RGV01-SURVEY"]</MudNavLink>
-            <MudNavLink Href="/4/4_4_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Android">@L["4.4.2.RGV02-SURVEY"]</MudNavLink>
+           @* <MudNavLink Href="/4/4_4_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Android">@L["4.4.2.RGV02-SURVEY"]</MudNavLink>*@
         </MudNavGroup>
         <MudNavGroup Title=@L["4.5-SRM-OPERATE"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Android">
             <MudNavLink Href="/4/4_5_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Android">@L["4.5.1.SRM-ENABLE"]</MudNavLink>
@@ -42,7 +42,7 @@
         </MudNavGroup>
     </MudNavGroup>
 
-    <MudNavGroup Title=@L["5.Report"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Report">
+    <MudNavGroup Title=@L["5.Report"] IconColor="Color.Primary" Icon="@Icons.Material.Filled.Report">
         <MudNavLink Href="/5/5_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.TableView">@L["5.1.Audit trail"]</MudNavLink>
         <MudNavGroup Title="@L["5.2-Inbound"]" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.TableView">
             <MudNavLink Href="/5/5_2_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.TableView">@L["5.2.1.Order Receive"]</MudNavLink>
@@ -52,7 +52,7 @@
             <MudNavLink Href="/5/5_3_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.TableView">@L["5.3.1.Inventory location"]</MudNavLink>
             <MudNavLink Href="/5/5_3_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.TableView">@L["5.3.2.Inventory summary"]</MudNavLink>
             <MudNavLink Href="/5/5_3_3" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.TableView">@L["5.3.3.Inventory aging"]</MudNavLink>
-            <MudNavLink Href="/5/5_3_4" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.TableView">@L["5.3.4.Inventory Customer"]</MudNavLink>
+           @* <MudNavLink Href="/5/5_3_4" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.TableView">@L["5.3.4.Inventory Customer"]</MudNavLink>*@
         </MudNavGroup>
         <MudNavGroup Title="@L["5.4-Outbound"]" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.TableView">
             <MudNavLink Href="/5/5_4_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.TableView">@L["5.4.1.Order picking"]</MudNavLink>
@@ -72,7 +72,7 @@
 
     </MudNavGroup>
 
-    <MudNavGroup Title=@L["6.Setup"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Settings">
+    <MudNavGroup Title=@L["6.Setup"] IconColor="Color.Primary" Icon="@Icons.Material.Filled.Settings">
         <MudNavLink Href="/6/6_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Settings">@L["SetMaterial"]</MudNavLink>
         <MudNavLink Href="/6/6_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Settings">@L["SetWorktype"]</MudNavLink>
         <MudNavLink Href="/6/6_3" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Settings">@L["SetStatus"]</MudNavLink>
@@ -80,13 +80,13 @@
         <MudNavLink Href="/6/6_5" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Settings">@L["SetStore"]</MudNavLink>
     </MudNavGroup>
 
-    <MudNavGroup Title=@L["7.Admin"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.AdminPanelSettings">
+    <MudNavGroup Title=@L["7.Admin"] IconColor="Color.Primary" Icon="@Icons.Material.Filled.AdminPanelSettings">
         <MudNavLink Href="/7/7_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.AdminPanelSettings">@L["AdmPrivileges"]</MudNavLink>
         <MudNavLink Href="/7/7_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.AdminPanelSettings">@L["AdmUser"]</MudNavLink>
         <MudNavLink Href="/7/7_3" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.AdminPanelSettings">@L["AdmGroup"]</MudNavLink>
     </MudNavGroup>
 
-    <MudNavGroup Title=@L["8.Help"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Help">
+    <MudNavGroup Title=@L["8.Help"] IconColor="Color.Primary" Icon="@Icons.Material.Filled.Help">
         <MudNavGroup Title="@L["HlpUserguide"]" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Help">
             <MudNavLink Href="/8/8_1_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Help">@L["HlpUser"]</MudNavLink>
             <MudNavLink Href="/8/8_1_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Help">@L["HlpAdmin"]</MudNavLink>
@@ -97,13 +97,23 @@
         <MudNavLink Href="/8/8_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Help">@L["HlpHelpdesk"]</MudNavLink>
     </MudNavGroup>
 
-    <MudNavGroup Title=@L["9.ERP"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.GroupWork">
+    <MudNavGroup Title="Maintenance" IconColor="Color.Primary" Icon="@Icons.Material.Filled.Assessment">
+             <MudNavLink Href="/6/6_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Assessment">Asset Register</MudNavLink>
+             <MudNavLink Href="/6/6_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Assessment">Preventative Maintenance</MudNavLink>
+             <MudNavLink Href="/6/6_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Assessment">History</MudNavLink>
+             <MudNavLink Href="/6/6_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Assessment">Control Files</MudNavLink>
+             <MudNavLink Href="/6/6_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Assessment">Reports</MudNavLink>
+             <MudNavLink Href="/6/6_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Assessment">Equipment Readings</MudNavLink>
+             <MudNavLink Href="/6/6_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.Assessment">Work Orders</MudNavLink>
+    </MudNavGroup>
+
+  @*  <MudNavGroup Title=@L["9.ERP"] IconColor="Color.Secondary" Icon="@Icons.Material.Filled.GroupWork">
         <MudNavLink Href="/9/9_1" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.GroupWork">@L["ErpBookingNote"]</MudNavLink>
         <MudNavLink Href="/9/9_2" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.GroupWork">@L["ErpDeliveryNote"]</MudNavLink>
         <MudNavLink Href="/9/9_3" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.GroupWork">@L["ErpItems"]</MudNavLink>
         <MudNavLink Href="/9/9_4" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.GroupWork">@L["ErpCustomers"]</MudNavLink>
         <MudNavLink Href="/9/9_5" Match="NavLinkMatch.Prefix" IconColor="Color.Secondary" Icon="@Icons.Material.Filled.GroupWork">@L["ErpPackeing"]</MudNavLink>
-    </MudNavGroup>
+    </MudNavGroup>*@
 </MudNavMenu>
 
 @code { string AlertAssignmentIcon { get; set; } = "<path d=\"M19,3A2,2 0 0,1 21,5V19A2,2 0 0,1 19,21H5A2,2 0 0,1 3,19V5A2,2 0 0,1 5,3H9.18C9.6,1.84 10.7,1 12,1C13.3,1 14.4,1.84 14.82,3H19M12,3A1,1 0 0,0 11,4A1,1 0 0,0 12,5A1,1 0 0,0 13,4A1,1 0 0,0 12,3M7,7V5H5V19H19V5H17V7H7M11,9H13V13.5H11V9M11,15H13V17H11V15Z\" />"; } 
